adiciona classe Display exibição de textos
